MILESTONE 2/creditsense/requirements.txt
streamlit>=1.32.0
fastapi>=0.111.0
uvicorn>=0.30.0
langgraph>=0.1.0
chromadb>=0.4.0
sentence-transformers
pdfplumber
python-docx
fpdf2
reportlab
httpx
pydantic>=2.0
python-dotenv
scikit-learn
joblib
typing-extensions
numpy
pandas
